feat(community): community home page shows subscribed posts, off-topic and feedback category grouped under "general".

+++ csc1106/src/main/java/webprogramming/csc1106/Repositories/PostRepo.java
@@ -7,7 +7,7 @@ import org.springframework.data.domain.Page;
 import webprogramming.csc1106.Entities.Post;
 import org.springframework.data.jpa.repository.Query;
 
-public interface PostRepo extends JpaRepository<Post, Integer>{
+public interface PostRepo extends JpaRepository<Post, Long>{
     Page<Post> findAllByTitleContainingOrContentContainingOrderByTimestampDesc(String title, String content, Pageable pageable);
     List<Post> findTop5ByOrderByTimestampDesc();
     Post findByPostID(long id);
